requirements.txt
argcomplete==2.0.0
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.94
botocore==1.27.94        
certifi==2022.6.15       
cffi==1.15.1
cfn-flip==1.3.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-rest-registration==0.7.2
django-stdimage==6.0.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
durationpy==0.5
hjson==3.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.1
oauthlib==3.2.0
pdfkit==1.0.0
Pillow==9.2.0
placebo==0.9.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tqdm==4.64.1
troposphere==4.1.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.11
Werkzeug==2.2.2
wsgi-request-logger==0.4.6
zappa==0.55.0